requirements.txt
alpaca-py==0.39.1
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
fastapi==0.115.5
idna==3.10
influxdb3-python==0.12.0
msgpack==1.1.0
numpy==2.2.4
pandas==2.2.3
pyarrow==19.0.1
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
reactivex==4.0.4
requests==2.32.3
six==1.17.0
sseclient-py==1.8.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.32.1
websockets==14.2
